requirements.txt
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.2.0
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.5
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
